# contracts/openapi/servicehub-v1.yaml
openapi: 3.1.0
info:
  title: ServiceHub 工单平台 API
  version: 1.0.0-draft.1
  description: |
    Vue 3 门户与 Spring Boot 单体服务的共享 v1 契约。所有业务 API 都由服务端执行对象级、流程动作和数据范围鉴权；UI 隐藏不是授权控制。
    附件与知识库文档必须先进入隔离区，完成服务端白名单校验和恶意文件扫描后才可下载、索引或发布。消息仅通过受控站内信和受管通道适配器投递；不得以通用回调、任意 URL 或客户端指定收件人替代。
    本版本不包含 AI、向量检索、自动生成、自动问答或自动处置能力。
    外部监控、CMDB、日志与 APM 仅通过受管连接器集成；本版本不实现自建监控平台、eBPF、自动发现 CMDB 或数据库运维平台。
servers:
  - url: /api/v1
    description: 同源网关路径
...
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: 在当前数据范围内查询工单
      description: 服务端忽略或拒绝越出当前授权范围的组织、队列、人员等筛选条件；不会依赖前端过滤。
      operationId: listTickets
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TicketType'
        - name: q
          in: query
          description: 关键词；服务端执行长度限制、转义和权限过滤。
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: 已按数据范围过滤的分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    post:
      tags: [Tickets]
      summary: 创建工单
      description: 以当前认证身份作为提单人。requesterId、组织、优先级、队列、状态和处理人均不得由客户端指定。
      operationId: createTicket
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: 工单已创建
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
              description: 新工单相对地址
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /tickets/{ticketId}:
    get:
      tags: [Tickets]
      summary: 获取单张工单
      description: 详情、评论、附件及关联信息需逐对象重新授权；无权访问时不得借由错误消息泄露对象存在性。
      operationId: getTicket
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 已授权的工单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tickets/{ticketId}/relations:
    get:
      tags: [Tickets]
      summary: 查询当前主体可见的关联工单
      description: |
        先校验当前工单，再对每一张关联目标重新执行对象读取授权；无权目标不得以编号、标题、数量或错误信息泄露存在性。
        关系不改变工单状态、审批、处理人或 SLA。
      operationId: listTicketRelations
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 当前主体可见的受控关系
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketRelation'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Tickets]
      summary: 建立受控工单关系
      description: |
        调用者必须具有源单修改权和目标单读取权。客户端仅提交目标编号和受控关系类型，不能提交目标摘要、组织范围、审批结论或状态。
        同一关系按自然键去重；`RELATED` 由服务端规范化为无方向关系。
      operationId: createTicketRelation
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketRelationRequest'
      responses:
        '201':
          description: 关系已创建或安全幂等确认
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketRelation'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tickets/{ticketId}/workflow:
    get:
      tags: [Workflow]
      summary: 获取当前授权主体的工单工作流概览
      description: 统一返回平台工作流投影、活动任务、协作人员快照、审批申请与服务端计算的可用动作。`controlledJumpActions` 仅会向服务经理或平台管理员返回，且只是显示能力；预演和执行仍在提交时重新执行对象授权、角色、审批、版本与节点校验。读取不授予写权限。
      operationId: getTicketWorkflowOverview
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 当前主体有权查看的工作流读模型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowOverviewReadModel'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tickets/{ticketId}/workflow/actions:
    post:
      tags: [Workflow]
      summary: 提交服务端受控的工单流程动作
      description: 客户端仅可提交动作意图和受限输入；状态、节点、候选人、组织范围、SLA 与通知收件人均由服务端重新计算。`CONTROLLED_JUMP_REQUEST` 只创建 Flowable 审批实例，不执行节点迁移。
      operationId: executeTicketWorkflowAction
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/WorkflowIfMatch'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowActionRequest'
      responses:
        '200':
          description: 动作已提交并返回最新工单投影
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /tickets/{ticketId}/workflow/approval-requests/{requestId}/decisions:
    post:
      tags: [Workflow]
      summary: 对受控跳转 Flowable 审批任务作出决策
      description: 仅服务经理或平台管理员候选角色可处理，且申请人不得自审。服务端先完成独立 Flowable 用户任务，再原子更新审批投影和审计；通过并不等于执行迁移。
      operationId: decideControlledJumpApproval
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/WorkflowApprovalRequestId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: 独立审批流程已完成，返回审批投影
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlledJumpApprovalProjection'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /tickets/{ticketId}/workflow/approval-requests/{requestId}/preflight:
    get:
      tags: [Workflow]
      summary: 只读预演已通过的受控跳转
      description: 仅服务经理或平台管理员可调用。验证审批结论、申请时的工单/流程版本、当前源节点与发布白名单，并返回任务处置、候选人、SLA 和通知重算计划。任何门禁失败均不改变 Flowable、工单、SLA 或通知。
      operationId: preflightControlledJumpExecution
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/WorkflowApprovalRequestId'
      responses:
        '200':
          description: 预演结论（`executable=false` 仍是正常受控响应）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlledJumpPreflight'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tickets/{ticketId}/workflow/approval-requests/{requestId}/execute:
    post:
      tags: [Workflow]
      summary: 执行已获批且预演通过的受控跳转
      description: 仅服务经理或平台管理员可调用。服务端以审批申请状态、申请时版本、对象授权和当前节点为门禁，先原子保留执行权，再在同一事务内迁移 Flowable、取消旧任务、创建新投影、重算 SLA/收件人、写入执行结果和审计。重复调用或版本变化返回冲突。
      operationId: executeApprovedControlledJump
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/WorkflowApprovalRequestId'
        - $ref: '#/components/parameters/WorkflowIfMatch'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: 已执行并返回最新工单投影
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SERVICEHUB_SESSION
      description: 由 Spring Boot 维护的 HttpOnly、Secure、SameSite 会话 Cookie；浏览器不得将访问令牌写入 localStorage。
    integrationCallbackSignature:
      type: apiKey
      in: header
      name: X-Integration-Signature
      description: 仅受管连接器回调使用的签名头。服务端必须基于受管 secretRef、原始请求体、时间戳和事件 ID 验签；该头不是用户会话或通用 API 密钥。
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 对所有非安全方法必传；令牌绑定当前会话。
      schema:
        type: string
        minLength: 20
        maxLength: 512
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        客户端为一次写入意图生成的 UUID。服务端将键绑定到认证主体、接口、目标工单/任务和规范化请求指纹；
        同一绑定在受控保留期内仅执行一次并返回原结果，使用相同键但不同指纹必须返回 IDEMPOTENCY_CONFLICT。
        键不可跨主体、工单、任务或操作复用，且服务端记录首次接收时间、结果和审计关联以防重放。
      schema:
        type: string
        format: uuid
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        pattern: '^TKT-[0-9]{8}-[0-9]{6}$'
    WorkflowApprovalRequestId:
      name: requestId
      in: path
      required: true
      description: 受控跳转审批申请 UUID；只用于定位，不能替代当前工单、审批候选资格或数据范围授权。
      schema:
        type: string
        format: uuid
    WorkflowIfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前工单版本的 HTTP 强实体标签；必须使用双引号包裹，例如 `"0"`。服务端以此执行乐观锁并拒绝过期页面或重复执行。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
        example: '"0"'
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationFailed:
      description: 请求字段或服务目录规则校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthenticated:
      description: 会话不存在、过期或 IAM 身份不可用
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: 已认证但不具备当前对象、动作或数据范围权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 对象不存在或按防枚举策略不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 幂等键冲突、状态冲突或乐观锁冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    RateLimited:
      description: 请求频率超过当前主体或来源限制
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ServiceUnavailable:
      description: 服务暂不可接收流量
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    TicketType:
      type: string
      enum: [INCIDENT, SERVICE_REQUEST, ACCESS_REQUEST, PROBLEM, CHANGE]
    TicketStatus:
      type: string
      enum: [DRAFT, SUBMITTED, PENDING_CLASSIFICATION, PENDING_ASSIGNMENT, PENDING_ACCEPTANCE, IN_PROGRESS, RESOLVED, PENDING_USER_FEEDBACK, CLOSED, CANCELLED, ON_HOLD]
...
    TicketPriority:
      type: string
      enum: [P1, P2, P3, P4]
...
    FormData:
      type: object
      description: |
        字段键仅限指定目录表单版本的 FormField.code；服务端拒绝未知、隐藏/无权、只读、类型不符、无效字典值及未满足联动的字段。
        该开放对象只为适配受控动态表单，不能承载状态、角色、人员、组织、队列、优先级、审批、路由、SLA 或任意流程变量。
      additionalProperties: true
...
    Tag:
      type: object
      additionalProperties: false
      required: [name, kind]
      properties:
        name:
          type: string
          pattern: '^#[^\\s#]{1,50}$'
          description: 自由标签必须以 # 开头；标准标签同样按此格式展示。
        kind:
          type: string
          enum: [STANDARD, FREE]
...
    TicketCreateRequest:
      type: object
      additionalProperties: false
      required: [serviceCatalogItemId, serviceCatalogFormVersion, type, title, description, structuredFields]
      x-server-enforced:
        - authenticated-requester-identity-snapshot
        - published-catalog-visibility-and-form-version-check
        - dynamic-form-field-type-linkage-and-dictionary-validation
        - tag-policy-and-normalization
        - service-side-routing-priority-approval-and-sla-calculation
      properties:
        serviceCatalogItemId:
          type: string
          maxLength: 64
          description: 服务目录项 ID；后端以此加载必填项、路由、审批与 SLA 规则。
        serviceCatalogFormVersion:
          type: integer
          minimum: 1
          description: 用户实际填写时看到的已发布表单版本。版本过期、目录不可见或表单摘要不一致时由服务端拒绝或要求刷新，不能回退至草稿。
        type:
          $ref: '#/components/schemas/TicketType'
        title:
          type: string
          minLength: 4
          maxLength: 200
        description:
          type: string
          maxLength: 16000
          description: 纯文本，或 `descriptionFormat=RICH_TEXT` 时的编辑器 HTML。服务端只存储白名单清洗后的 HTML 与派生纯文本摘要。
        descriptionFormat:
          type: string
          enum: [PLAIN_TEXT, RICH_TEXT]
          default: PLAIN_TEXT
          description: RICH_TEXT 仅允许平台规定的 TipTap 子集；脚本、样式、图片、嵌入内容与非 HTTP(S) 链接均会被拒绝或移除。
        structuredFields:
          $ref: '#/components/schemas/FormData'
        tags:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/Tag'
        relatedConfigurationItemIds:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 128
...
    WorkflowActionRequest:
      type: object
      additionalProperties: false
      required: [action]
      x-server-enforced:
        - ticket-object-authorization
        - server-calculated-state-node-assignment-sla-and-notification-recipients
        - ticket-optimistic-lock
        - immutable-audit-event
      properties:
        action:
          type: string
          enum: [CLASSIFY, ASSIGN, ACCEPT, START_PROCESSING, REQUEST_USER_FEEDBACK, RESOLVE, CLOSE, REOPEN, CANCEL, HOLD, RESUME, ESCALATE, CLAIM, TRANSFER, ADD_COHANDLER, HANDOVER, INTERNAL_COMMENT, CONTROLLED_JUMP_REQUEST]
        targetIamUserId:
          type: string
          pattern: '^[A-Za-z0-9._:-]{1,128}$'
          description: 仅作为转办、协办、交接或分派候选人；服务端重新校验 IAM 有效性、角色、班次与数据范围。
        comment:
          type: string
          maxLength: 2000
        reason:
          type: string
          maxLength: 1000
        targetNode:
          type: string
          enum: [classify, assign, accept, processing, user_feedback, closure]
          description: 仅受控跳转申请使用，表示目标节点意图，不会直接改变 BPMN 节点。
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [decision, reason]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
          minLength: 5
          maxLength: 1000
...
    WorkflowOverviewReadModel:
      type: object
      additionalProperties: false
      required: [instance, tasks, comments, availableActions, events, participants, approvalRequests, approvalDecisions, controlledJumpActions]
      properties:
        instance:
          $ref: '#/components/schemas/WorkflowInstanceProjection'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowTaskProjection'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowCommentProjection'
        availableActions:
          type: array
          description: 仅当前读取时的展示提示；写入时服务端仍重新鉴权。
          items:
            $ref: '#/components/schemas/WorkflowAvailableAction'
        events:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowEventProjection'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowParticipantProjection'
        approvalRequests:
          type: array
          items:
            $ref: '#/components/schemas/ControlledJumpApprovalProjection'
        approvalDecisions:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalDecisionRecord'
        controlledJumpActions:
          type: array
          description: 仅服务经理或平台管理员可见的服务端计算管理能力；不是授权凭据。
          items:
            $ref: '#/components/schemas/ControlledJumpManagementAction'
...
    ControlledJumpApprovalProjection:
      type: object
      additionalProperties: false
      required: [id, ticketId, applicantIamUserId, sourceNode, targetNode, reason, status, createdAt, approvalPolicy]
      properties:
        id:
          type: string
          format: uuid
        ticketId:
          $ref: '#/components/schemas/TicketIdValue'
        applicantIamUserId:
          type: string
          maxLength: 128
        sourceNode:
          type: string
          maxLength: 64
        targetNode:
          type: string
          maxLength: 64
        reason:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [PENDING_APPROVAL, APPROVED, REJECTED, EXECUTING, EXECUTED]
        createdAt:
          type: string
          format: date-time
        approverIamUserId:
          type: string
          maxLength: 128
        decisionReason:
          type: string
          maxLength: 1000
        decidedAt:
          type: string
          format: date-time
        approvalEngineInstanceId:
          type: string
          maxLength: 128
        sourceTicketVersion:
          type: integer
          minimum: 0
        sourceWorkflowVersion:
          type: integer
          minimum: 0
        executorIamUserId:
          type: string
          maxLength: 128
        executionStartedAt:
          type: string
          format: date-time
        executedAt:
          type: string
          format: date-time
        executedFromNode:
          type: string
          maxLength: 64
        executedToNode:
          type: string
          maxLength: 64
        executionFailureReason:
          type: string
          maxLength: 160
        approvalPolicy:
          $ref: '#/components/schemas/ApprovalPolicySnapshot'
...
    ControlledJumpPreflight:
      type: object
      additionalProperties: false
      required: [executable, blockingReasons, currentTaskDisposition, targetCandidateRole, candidateResolution, candidateRecalculationRequired, slaImpact, notificationImpact]
      properties:
        executable:
          type: boolean
        blockingReasons:
          type: array
          items:
            type: string
            maxLength: 200
        currentTaskDisposition:
          type: string
          maxLength: 128
        targetCandidateRole:
          type: string
          maxLength: 128
        candidateResolution:
          type: string
          maxLength: 128
        candidateRecalculationRequired:
          type: boolean
        slaImpact:
          type: string
          maxLength: 128
        notificationImpact:
          type: string
          maxLength: 128
...
    TicketIdValue:
      type: string
      pattern: '^TKT-[0-9]{8}-[0-9]{6}$'
...
    Ticket:
      type: object
      additionalProperties: false
      required: [id, type, status, priority, title, requester, serviceCatalogItem, createdAt, version]
      properties:
        id:
          type: string
          pattern: '^TKT-[0-9]{8}-[0-9]{6}$'
        type:
          $ref: '#/components/schemas/TicketType'
        status:
          $ref: '#/components/schemas/TicketStatus'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 4000
          description: 服务端从原始输入或已清洗富文本派生的纯文本摘要；用于检索和非富文本兼容展示。
        descriptionFormat:
          type: string
          enum: [PLAIN_TEXT, RICH_TEXT]
        descriptionHtml:
          type: string
          maxLength: 12000
          description: 仅当 descriptionFormat=RICH_TEXT 时出现，且已由服务端白名单清洗。客户端不得将任意其他字段以 HTML 渲染。
        requester:
          $ref: '#/components/schemas/IdentitySnapshot'
        assignee:
          $ref: '#/components/schemas/IdentitySnapshot'
        serviceCatalogItem:
          $ref: '#/components/schemas/ServiceCatalogSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 0
          description: 乐观锁版本；写操作须在后续动作契约中携带并校验。
    TicketRelationType:
      type: string
      enum: [RELATED, DUPLICATE_OF, PARENT_OF, PROBLEM_REFERENCE, CHANGE_REFERENCE]
    TicketRelationRequest:
      type: object
      additionalProperties: false
      required: [targetTicketId, relationType]
      properties:
        targetTicketId:
          type: string
          pattern: '^TKT-[0-9]{8}-[0-9]{6}$'
        relationType:
          $ref: '#/components/schemas/TicketRelationType'
    TicketRelation:
      type: object
      additionalProperties: false
      required: [relationType, direction, relatedTicket, createdByIamUserId, createdAt]
      properties:
        relationType:
          $ref: '#/components/schemas/TicketRelationType'
        direction:
          type: string
          enum: [OUTBOUND, INBOUND]
          description: 相对于当前查询工单的关系方向；RELATED 的方向仅用于审计展示。
        relatedTicket:
          $ref: '#/components/schemas/TicketRelationSummary'
        createdByIamUserId:
          type: string
          maxLength: 128
        createdAt:
          type: string
          format: date-time
...
    ServiceCatalogSummary:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 200
    TicketPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    ApiError:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: 稳定问题类型，例如 /problems/validation-failed。
        title:
          type: string
          maxLength: 160
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum: [VALIDATION_FAILED, DYNAMIC_FORM_VALIDATION_FAILED, CATALOG_NOT_AVAILABLE, RULE_MATCH_NOT_ALLOWED, ATTACHMENT_REJECTED, ATTACHMENT_NOT_AVAILABLE, ATTACHMENT_SCAN_UNAVAILABLE, KNOWLEDGE_IMPORT_REJECTED, KNOWLEDGE_DOCUMENT_NOT_AVAILABLE, KNOWLEDGE_PUBLICATION_NOT_ALLOWED, INTEGRATION_CONFIGURATION_NOT_ALLOWED, INTEGRATION_CALLBACK_REJECTED, INTEGRATION_REPLAY_REJECTED, CONFIGURATION_ITEM_NOT_AVAILABLE, ALERT_NOT_AVAILABLE, OBSERVABILITY_LINK_NOT_AVAILABLE, NOTIFICATION_NOT_AVAILABLE, NOTIFICATION_RECIPIENT_NOT_ALLOWED, NOTIFICATION_TEMPLATE_REJECTED, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, INVALID_WORKFLOW_ACTION, INVALID_PROCESS_NODE, APPROVAL_REQUIRED, RATE_LIMITED, SERVICE_UNAVAILABLE, INTERNAL_ERROR]
        detail:
          type: string
          maxLength: 500
          description: 面向用户的安全摘要；禁止包含 SQL、堆栈、令牌、权限规则或对象敏感信息。
        requestId:
          type: string
          maxLength: 64
        fieldErrors:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
                maxLength: 200
              code:
                type: string
                maxLength: 80
              message:
                type: string
                maxLength: 300
security:
  - sessionCookie: []